Add common Vue components and restructure home page integration
