+++ common/client/views/Footer.vue
