+++ home/web/Controllers/HomeController.cs
@@ -19,11 +19,6 @@ namespace Nagiyu.Web.Controllers
             return View();
         }
 
-        public IActionResult Privacy()
-        {
-            return View();
-        }
-
         [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
         public IActionResult Error()
         {
