+++ home/web/Views/Home/Index.cshtml
@@ -5,5 +5,5 @@
 <div id="vue-home" class="content"></div>
 
 @section Scripts {
-    @* <script type="module" src="~/js/dist/home.js"></script> *@
+    <script type="module" src="~/js/dist/home.js"></script>
 }
