+++ home/web/Views/Shared/_Layout.cshtml
@@ -38,7 +38,7 @@
 
     <footer id="vue-footer" class="footer"></footer>
 
-    <script type="module" src="~/js/dist/home.js"></script>
+    <script type="module" src="~/js/dist/common.js"></script>
     @await RenderSectionAsync("Scripts", required: false)
 </body>
